create pytest_indexd.indexd_client_fixture
