+++ .secrets.baseline
@@ -169,7 +169,7 @@
         "filename": "pytest_indexd/__init__.py",
         "hashed_secret": "d033e22ae348aeb5660fc2140aec35850c4da997",
         "is_verified": false,
-        "line_number": 81,
+        "line_number": 80,
         "is_secret": false
       }
     ],
@@ -208,5 +208,5 @@
       }
     ]
   },
-  "generated_at": "2023-01-10T15:47:58Z"
+  "generated_at": "2023-01-10T15:49:47Z"
 }
