+++ pytest_indexd/plugin.py
@@ -4,8 +4,8 @@ from pytest_indexd import (
     create_indexd_tables,
     index_driver,
     indexd_admin_user,
-    indexd_client,
     pg_url,
     postgresql_server_indexd,
 )
+from pytest_indexd.indexd_client_fixture import indexd_client
 from pytest_indexd.indexd_server_fixture import indexd_server
